chore: move viewableAccounts & liveAccounts to Dashboard.svelte to expose globally
